refactor(dashboard): extract shared utilities — findKeeper, isOfflineStatus, formatTokens
- Move formatTokens from keeper-detail-panels.ts to lib/format-number.ts (#4204)
- Extract findKeeper() to lib/keeper-utils.ts, replace 5 duplicate definitions (#4202)
- Extract isOfflineStatus() to lib/status-utils.ts, replace 7 inline checks (#4203)
- Remove dead normalizeStatus() from tool-audit.ts
- Remove unused keepers/Keeper imports from 4 files

Closes #4204, closes #4202, closes #4203
